base: provide default retrieval impls
Implement default implementations for entity-specific repositories'
methods on retrieval operations. Streaming implementations aren't
included. This allows users to have to implement less to get a working
cache. The downside is that this can be inefficient due to being two
cache requests, when a backend may be able to do an operation in one, so
it is still recommended to implement these manually.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>